+++ crates/runc/src/lib.rs
@@ -38,7 +38,6 @@
 use std::fmt::{self, Display};
 use std::path::{Path, PathBuf};
 use std::process::ExitStatus;
-use std::time::Duration;
 
 use oci_spec::runtime::{Linux, Process};
 
@@ -98,177 +97,9 @@ impl Default for LogFormat {
     }
 }
 
-/// Global options builder for the runc binary.
-///
-/// These options will be passed for all subsequent runc calls.
-/// See <https://github.com/opencontainers/runc/blob/main/man/runc.8.md#global-options>
-#[derive(Debug, Default)]
-pub struct ConfigBuilder {
-    /// Override the name of the runc binary. If [`None`], `runc` is used.
-    command: Option<PathBuf>,
-    /// Path to root directory of container rootfs.
-    root: Option<PathBuf>,
-    /// Debug logging.
-    ///
-    /// If true, debug level logs are emitted.
-    debug: bool,
-    /// Path to log file.
-    log: Option<PathBuf>,
-    /// Log format to use.
-    log_format: LogFormat,
-    /// Set process group ID (gpid).
-    set_pgid: bool,
-    /// Use systemd cgroup.
-    systemd_cgroup: bool,
-    /// Whether to use rootless mode.
-    ///
-    /// If [`None`], `auto` settings is used.
-    /// Note that "auto" is different from explicit "true" or "false".
-    rootless: Option<bool>,
-    /// Timeout settings for runc command.
-    ///
-    /// Default is 5 seconds.
-    /// This will be used only in AsyncClient.
-    timeout: Duration,
-}
-
 /// A shortcut to create `runc` global options builder.
-pub fn builder() -> ConfigBuilder {
-    ConfigBuilder::default()
-}
-
-impl ConfigBuilder {
-    /// Create new config builder with no options.
-    pub fn new() -> Self {
-        Default::default()
-    }
-
-    pub fn command(mut self, command: impl AsRef<Path>) -> Self {
-        self.command = Some(command.as_ref().to_path_buf());
-        self
-    }
-
-    /// Set the root directory to store containers' state.
-    ///
-    /// The path should be located on tmpfs.
-    /// Default is `/run/runc`, or `$XDG_RUNTIME_DIR/runc` for rootless containers.
-    pub fn root(mut self, root: impl AsRef<Path>) -> Self {
-        self.root = Some(root.as_ref().to_path_buf());
-        self
-    }
-
-    /// Enable debug logging.
-    pub fn debug(mut self, debug: bool) -> Self {
-        self.debug = debug;
-        self
-    }
-
-    /// Set the log destination to path.
-    ///
-    /// The default is to log to stderr.
-    pub fn log(&mut self, log: impl AsRef<Path>) -> &mut Self {
-        self.log = Some(log.as_ref().to_path_buf());
-        self
-    }
-
-    /// Set the log format (default is text).
-    pub fn log_format(mut self, log_format: LogFormat) -> Self {
-        self.log_format = log_format;
-        self
-    }
-
-    /// Set the log format to JSON.
-    pub fn log_json(self) -> Self {
-        self.log_format(LogFormat::Json)
-    }
-
-    /// Set the log format to TEXT.
-    pub fn log_text(self) -> Self {
-        self.log_format(LogFormat::Text)
-    }
-
-    /// Enable systemd cgroup support.
-    ///
-    /// If this is set, the container spec (`config.json`) is expected to have `cgroupsPath` value in
-    // the `slice:prefix:name` form (e.g. `system.slice:runc:434234`).
-    pub fn systemd_cgroup(mut self, systemd_cgroup: bool) -> Self {
-        self.systemd_cgroup = systemd_cgroup;
-        self
-    }
-
-    /// Enable or disable rootless mode.
-    ///
-    // Default is auto, meaning to auto-detect whether rootless should be enabled.
-    pub fn rootless(mut self, rootless: bool) -> Self {
-        self.rootless = Some(rootless);
-        self
-    }
-
-    /// Set rootless mode to auto.
-    pub fn rootless_auto(mut self) -> Self {
-        self.rootless = None;
-        self
-    }
-
-    pub fn set_pgid(mut self, set_pgid: bool) -> Self {
-        self.set_pgid = set_pgid;
-        self
-    }
-
-    pub fn timeout(&mut self, millis: u64) -> &mut Self {
-        self.timeout = Duration::from_millis(millis);
-        self
-    }
-
-    fn build_runc(self) -> Result<Runc> {
-        let path = self
-            .command
-            .clone()
-            .unwrap_or_else(|| PathBuf::from("runc"));
-
-        let command = utils::binary_path(path).ok_or(Error::NotFound)?;
-
-        let mut args = Vec::new();
-
-        // --root path : Set the root directory to store containers' state.
-        if let Some(root) = self.root {
-            args.push("--root".into());
-            args.push(utils::abs_string(root)?);
-        }
-
-        // --debug : Enable debug logging.
-        if self.debug {
-            args.push("--debug".into());
-        }
-
-        // --log path : Set the log destination to path. The default is to log to stderr.
-        if let Some(log_path) = self.log {
-            args.push("--log".into());
-            args.push(utils::abs_string(log_path)?);
-        }
-
-        // --log-format text|json : Set the log format (default is text).
-        args.push("--log-format".into());
-        args.push(self.log_format.to_string());
-
-        // --systemd-cgroup : Enable systemd cgroup support.
-        if self.systemd_cgroup {
-            args.push("--systemd-cgroup".into());
-        }
-
-        // --rootless true|false|auto : Enable or disable rootless mode.
-        if let Some(mode) = self.rootless {
-            let arg = format!("--rootless={}", mode);
-            args.push(arg);
-        }
-
-        Ok(Runc { command, args })
-    }
-
-    pub fn build(self) -> Result<Runc> {
-        let runc = self.build_runc()?;
-        Ok(runc)
-    }
+pub fn builder() -> GlobalOpts {
+    GlobalOpts::default()
 }
 
 #[cfg(not(feature = "async"))]
@@ -304,6 +135,7 @@ impl Runc {
     }
 }
 
+//<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<
 #[cfg(not(feature = "async"))]
 impl Runc {
     pub fn checkpoint(&self) -> Result<()> {
@@ -601,7 +433,7 @@ impl Runc {
     }
 
     /// Return an event stream of container notifications
-    pub async fn events(&self, _id: &str, _interval: &Duration) -> Result<()> {
+    pub async fn events(&self, _id: &str, _interval: &std::time::Duration) -> Result<()> {
         Err(Error::Unimplemented("events".to_string()))
     }
 
@@ -728,6 +560,7 @@ impl Runc {
         Ok(())
     }
 }
+//>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
 
 #[cfg(test)]
 #[cfg(all(target_os = "linux", not(feature = "async")))]
@@ -735,14 +568,14 @@ mod tests {
     use super::*;
 
     fn ok_client() -> Runc {
-        ConfigBuilder::new()
+        GlobalOpts::new()
             .command("/bin/true")
             .build()
             .expect("unable to create runc instance")
     }
 
     fn fail_client() -> Runc {
-        ConfigBuilder::new()
+        GlobalOpts::new()
             .command("/bin/false")
             .build()
             .expect("unable to create runc instance")
@@ -875,14 +708,14 @@ mod tests {
     use super::*;
 
     fn ok_client() -> Runc {
-        ConfigBuilder::new()
+        GlobalOpts::new()
             .command("/bin/true")
             .build()
             .expect("unable to create runc instance")
     }
 
     fn fail_client() -> Runc {
-        ConfigBuilder::new()
+        GlobalOpts::new()
             .command("/bin/false")
             .build()
             .expect("unable to create runc instance")
